move ported code back to legacy
