--- gato/components/__init__.py
@@ -1,1 +1,0 @@
-from .components import *
