--- gato/multipart/__init__.py
@@ -1,2 +1,0 @@
-from .parser import *
-from .containers import *
